Update dashtrends.php
There's a problem with the daylight saving (DST) to show graphical information and calculations. In my country (Brazil) the DST was this month.
After the date 18 - october - 2015 all informations was shown as zero. These 2 lines correct this bug, it identifies the "1" on date hour field, correct your value, and after, get the timestamp again.